move fmap need/uneed into predict layer so timeout threads don't access released memory

+++ libclamav/predict.h
@@ -17,7 +17,7 @@ typedef struct PredictionResult_t {
 #define PREDICT_VIRNAME "Request_Inspection.AE"
 
 // callbacks into aescan: do the prediction, dispose of the results, and log things
-typedef PredictionResult* (*Predict_t)(const char *filename, const void *buf, uint32_t len);
+typedef PredictionResult* (*Predict_t)(const char *filename, const void *buf, size_t len);
 typedef void (*DisposePredictionResult_t)(PredictionResult* result);
 typedef void (*LogPredict_t)(const char *level, const char *msg);
